+++ src/momo_args.h
@@ -18,6 +18,9 @@ struct MomoArgs {
   bool no_google_stun = false;
   bool no_video_device = false;
   bool no_audio_device = false;
+#if defined(USE_FAKE_CAPTURE_DEVICE)
+  bool fake_capture_device = false;
+#endif
   bool force_i420 = false;
   // Jetson の場合だけデフォルト true
 #if defined(USE_JETSON_ENCODER)
@@ -30,6 +33,8 @@ struct MomoArgs {
   // use_libcamera == true の場合だけ使える。
   // sora_video_codec_type == "H264" かつ sora_simulcast == false の場合だけしか機能しない。
   bool use_libcamera_native = false;
+  // libcamera のコントロール設定。key value の形式で指定する。
+  std::vector<std::pair<std::string, std::string>> libcamera_controls;
   std::string video_device = "";
   std::string resolution = "VGA";
   int framerate = 30;